split media types
