+++ common/src/index.ts
@@ -19,10 +19,10 @@ import {
   BlogViewSchema,
   CreateBlog,
   CreateBlogSchema,
-  ReplaceBlog,
-  ReplaceBlogSchema,
   ModifyBlog,
   ModifyBlogSchema,
+  ReplaceBlog,
+  ReplaceBlogSchema,
 } from "./objects/blogs";
 import {
   CreateEvent,
@@ -31,10 +31,10 @@ import {
   EventSchema,
   FilterEvent,
   FilterEventSchema,
-  ReplaceEvent,
-  ReplaceEventSchema,
   ModifyEvent,
   ModifyEventSchema,
+  ReplaceEvent,
+  ReplaceEventSchema,
 } from "./objects/events";
 import {
   PaginatedResponse,
@@ -61,69 +61,75 @@ import {
 import {
   CreatePrice,
   CreatePriceSchema,
+  ModifyPrice,
+  ModifyPriceSchema,
   Price,
   PriceSchema,
   PriceView,
   PriceViewSchema,
   ReplacePrice,
   ReplacePriceSchema,
-  ModifyPrice,
-  ModifyPriceSchema,
 } from "./objects/prices";
 import {
   CreateProduction,
   CreateProductionSchema,
   FilterProduction,
   FilterProductionSchema,
+  ModifyProduction,
+  ModifyProductionSchema,
   Production,
   ProductionSchema,
   ProductionView,
   ProductionViewSchema,
   ReplaceProduction,
   ReplaceProductionSchema,
-  ModifyProduction,
-  ModifyProductionSchema,
 } from "./objects/productions";
 import {
   CreateTag,
   CreateTagSchema,
+  ModifyTag,
+  ModifyTagSchema,
   Tag,
   TagSchema,
   TagView,
   TagViewSchema,
-  ModifyTag,
-  ModifyTagSchema,
 } from "./objects/tags";
 import {
-  CreateMediaCrop,
-  CreateMediaCropSchema,
   CreateMediaGallery,
   CreateMediaGallerySchema,
-  CreateMediaItem,
-  CreateMediaItemSchema,
-  CropName,
-  CropNameEnum,
-  ItemPosition,
-  ItemPositionEnum,
-  MediaCrop,
-  MediaCropSchema,
+  GalleryType,
+  GalleryTypeEnum,
   MediaGallery,
   MediaGallerySchema,
+  ModifyMediaGallery,
+  ModifyMediaGallerySchema,
+} from "./objects/media_gallery";
+import {
+  CreateMediaItem,
+  CreateMediaItemSchema,
+  ItemPosition,
+  ItemPositionEnum,
   MediaItem,
   MediaItemSchema,
   MediaItemView,
   MediaItemViewSchema,
-  ModifyMediaCrop,
-  ModifyMediaCropSchema,
-  ModifyMediaGallery,
-  ModifyMediaGallerySchema,
   ModifyMediaItem,
   ModifyMediaItemSchema,
-  ReplaceMediaCrop,
-  ReplaceMediaCropSchema,
   ReplaceMediaItem,
   ReplaceMediaItemSchema,
-} from "./objects/media";
+} from "./objects/media_item";
+import {
+  CreateMediaCrop,
+  CreateMediaCropSchema,
+  CropName,
+  CropNameEnum,
+  MediaCrop,
+  MediaCropSchema,
+  ModifyMediaCrop,
+  ModifyMediaCropSchema,
+  ReplaceMediaCrop,
+  ReplaceMediaCropSchema,
+} from "./objects/media_crop";
 
 // list of all exports: (this way only need to import this file.)
 export {
@@ -183,7 +189,7 @@ export {
   DEFAULT_LANGUAGE,
   SUPPORTED_LANGUAGES,
 
-  // Media Schemas. (media.ts)
+  // Media Schemas. (media_*.ts)
   MediaGallerySchema,
   MediaItemSchema,
   MediaItemViewSchema,
@@ -198,6 +204,7 @@ export {
   ReplaceMediaItemSchema,
   CropNameEnum,
   ItemPositionEnum,
+  GalleryTypeEnum,
 };
 export type {
   // Production Types. (productions.ts)
@@ -256,7 +263,7 @@ export type {
   Language,
   LanguageQuery,
 
-  // Media Types. (media.ts)
+  // Media Types. (media_*.ts)
   MediaGallery,
   MediaItem,
   MediaItemView,
@@ -271,4 +278,5 @@ export type {
   ReplaceMediaItem,
   CropName,
   ItemPosition,
+  GalleryType,
 };
